# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: go-auth
  version: v1
  license:
    name: MIT
    identifier: MIT
  description: |
    A self-hosted authentication service. Five endpoints; every request and
    response body is JSON. No endpoint takes a credential in a header.

    The `POST` endpoints require `Content-Type: application/json`
    (parameters such as `; charset=utf-8` are ignored), else `415`. A body is
    decoded only up to its first 64 KiB, else `413`. When the operator turns
    rate limiting on, register, login and refresh answer `429` with
    `Retry-After` once a client address has used up its allowance; verify is
    never limited.

    A path no endpoint has is answered `404` with code `route_not_found`, in the
    usual error shape. A path not in canonical form (`//v1/auth/login`, or one
    with `..` segments) gets a `307` redirect from Go's `http.ServeMux`, not a
    JSON error.

    Branch on `error.code`, never on `message`. Treat a code you do not
    recognise as a generic failure: new codes can be added within a major
    version.

    The rules this file cannot express (the tolerances of the request-body
    decoder, refresh token reuse detection, the deliberately vague errors,
    caching the JWKS) are in `docs/api/reference.md`, `docs/api/errors.md` and
    `docs/api/token-verification.md`. The code is the source of truth: where
    this file or those documents disagree with it, the code is right and both
    are fixed.
servers:
...
security: []
paths:
...
  /v1/auth/refresh:
    post:
      operationId: refresh
      summary: Exchange a refresh token for a new pair
      description: |
        Exchanges a valid refresh token for a new access token and a new
        refresh token. The token sent is spent: replace your stored copy with
        the new one, since sending the old one again revokes the whole session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            examples:
              refresh:
                value:
                  refresh_token: kZ8m2Q1nR7yTxV3bC0dEfGhIjKlMnOpQrStUvWxYz01
      responses:
        '200':
          description: The token was valid and is now spent; here is the new pair.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokensResponse'
              examples:
                tokens:
                  value:
                    access_token:
                      value: eyJhbGciOiJFZERTQSIsImtpZCI6IjRxNi4uLiIsInR5cCI6IkpXVCJ9...
                      expires_at: '2026-09-08T12:30:00Z'
                      expires_in: 1799
                    refresh_token:
                      value: kZ8m2Q1nR7yTxV3bC0dEfGhIjKlMnOpQrStUvWxYz01
                      expires_at: '2026-09-15T12:00:00Z'
                      expires_in: 604800
        '400':
          $ref: '#/components/responses/InvalidJSONBody'
        '401':
          description: |
            The refresh token cannot be used. Every failure gets this one
            response: a token that never existed, one that expired, one already
            used, and one whose session was revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_token:
                  value:
                    error:
                      code: invalid_token
                      message: Not authorized.
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '413':
          $ref: '#/components/responses/RequestBodyTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: |
            The `refresh_token.value` from the last login or refresh. A missing
            field is read as `""`.
...
    Token:
      type: object
      required: [value, expires_at, expires_in]
      properties:
        value:
          type: string
        expires_at:
          type: string
          format: date-time
          description: |
            RFC 3339, in UTC, to the whole second. A fraction of a second is
            dropped, so it is never later than the real expiry.
        expires_in:
          type: integer
          format: int64
          description: |
            Whole seconds from the moment the token was issued to its expiry,
            rounded down. Set by `ACCESS_TOKEN_TTL` or `REFRESH_TOKEN_TTL`.
    TokensResponse:
      type: object
      required: [access_token, refresh_token]
      properties:
        access_token:
          $ref: '#/components/schemas/Token'
          description: |
            Its value is a JWT signed with EdDSA. Verify it with the keys from
            `/.well-known/jwks.json`; see `docs/api/token-verification.md`.
        refresh_token:
          $ref: '#/components/schemas/Token'
          description: |
            Its value is 32 random bytes in base64url without padding: not a
            JWT, and opaque outside this service. Keep it secret, and never put
            it in a URL.
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorData'
    ErrorData:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            A stable identifier for the failure. Branch on this. The catalog is
            in `docs/api/errors.md`; treat a code you do not recognise as a
            generic failure.
        message:
          type: string
          description: |
            Text for humans, set by the error's class rather than by its code.
            Do not parse it.
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
          description: |
            Present only when `code` is `validation_failed`: every field that
            failed. A field can appear more than once.
    FieldError:
      type: object
      required: [field, code, details]
      properties:
        field:
          type: string
          description: The request field that failed, named as in the JSON body.
        code:
          type: string
          description: |
            The rule that was broken: `too_short`, `too_long` or
            `invalid_characters` today.
        details:
          type: object
          description: |
            The rule's parameters, shaped by its code: `{min, unit}` for
            `too_short`, `{max, unit}` for `too_long`, `{}` for
            `invalid_characters`.
          properties:
            min:
              type: integer
            max:
              type: integer
            unit:
              type: string
              description: How the length was counted, `code_point` or `byte`.
  responses:
    InvalidJSONBody:
      description: |
        The body is not a JSON object with fields of the expected types. That
        includes an empty body and invalid UTF-8.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid_json_body:
              value:
                error:
                  code: invalid_json_body
                  message: Request body is not valid JSON.
    MethodNotAllowed:
      description: |
        The path exists but does not accept the method used: the answer to a
        different method on this path, not to the one documented here.
      headers:
        Allow:
          required: true
          description: |
            The accepted methods: `POST` on the auth endpoints, `GET, HEAD` on
            the JWKS.
          schema:
            type: string
          example: POST
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            method_not_allowed:
              value:
                error:
                  code: method_not_allowed
                  message: Method not allowed for this route.
    RequestBodyTooLarge:
      description: The body's first JSON value does not end within 64 KiB.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            request_body_too_large:
              value:
                error:
                  code: request_body_too_large
                  message: Request body is too large.
    UnsupportedMediaType:
      description: The request's `Content-Type` is not `application/json`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unsupported_media_type:
              value:
                error:
                  code: unsupported_media_type
                  message: Content-Type must be application/json.
    TooManyRequests:
      description: |
        Rate limiting is on, and this client address used up the endpoint's
        allowance. Nothing was read or run: send the same request again after
        `Retry-After`, on `/v1/auth/refresh` with the same token.
      headers:
        Retry-After:
          required: true
          description: The seconds to wait before sending the request again.
          schema:
            type: integer
            minimum: 1
          example: 20
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            too_many_requests:
              value:
                error:
                  code: too_many_requests
                  message: Too many requests.
    InternalServerError:
      description: An unexpected failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            internal_server_error:
              value:
                error:
                  code: internal_server_error
                  message: An internal error occurred.